+++ src/cross-origin-isolation.ts
@@ -1,0 +1,17 @@
+/**
+ * @license
+ * Copyright 2024 Google LLC
+ * SPDX-License-Identifier: BSD-3-Clause
+ */
+import {Middleware} from 'koa';
+
+// Enable cross-origin isolation for more precise timers:
+// https://developer.chrome.com/blog/cross-origin-isolated-hr-timers/
+export function crossOriginIsolation(): Middleware {
+  // Based on https://github.com/fishel-feng/koa-isolated
+  return async function isolated(ctx, next) {
+    ctx.set('Cross-Origin-Opener-Policy', 'same-origin');
+    ctx.set('Cross-Origin-Embedder-Policy', 'require-corp');
+    await next();
+  };
+}